moved const.go to discovery.go
